Refactor: extract to method
This is to allow easier testing
