+++ server/services/location.js
@@ -1,8 +1,9 @@
 const { bingKeyLocation, bingUrl } = require('../config')
-const { getJson, addBufferToBbox, formatName } = require('../util')
-const floodServices = require('./flood')
+const { getJson } = require('../util')
 const util = require('util')
+const bingResultsParser = require('./lib/bing-results-parser')
 const LocationSearchError = require('../location-search-error')
+const floodServices = require('./flood')
 
 async function find (location) {
   const query = encodeURIComponent(location)
@@ -32,48 +33,7 @@ async function find (location) {
     throw new LocationSearchError('Invalid geocode results (no resourceSets)')
   }
 
-  const set = bingData.resourceSets[0]
-  if (set.estimatedTotal === 0) {
-    return []
-  }
-
-  const data = set.resources[0]
-
-  if (data.confidence.toLowerCase() === 'low') {
-    return []
-  }
-
-  const {
-    bbox,
-    point: { coordinates: center }
-  } = data
-
-  const name = formatName(data.name, data.entityType)
-
-  // Reverse as Bing returns as [y (lat), x (long)]
-  bbox.reverse()
-  center.reverse()
-
-  const isUK = data.address.countryRegionIso2 === 'GB'
-  const isScotlandOrNorthernIreland = isUK &&
-  (data.address.adminDistrict === 'Northern Ireland' || data.address.adminDistrict === 'Scotland')
-
-  const isEngland = await floodServices.getIsEngland(center[0], center[1])
-
-  // add on 2000m buffer to place.bbox for warnings and alerts search
-  const bbox2k = addBufferToBbox(bbox, 2000)
-  // add on 10000m buffer to place.bbox for stations search
-  const bbox10k = addBufferToBbox(bbox, 10000)
-
-  return [{
-    name,
-    center,
-    bbox2k,
-    bbox10k,
-    isEngland,
-    isUK,
-    isScotlandOrNorthernIreland
-  }]
+  return await bingResultsParser(bingData, floodServices.getIsEngland)
 }
 
 module.exports = { find }
